fix: blank problems were not being edited in a modal
