+++ src/editors/containers/ProblemEditor/components/SelectTypeModal/index.test.tsx
@@ -6,6 +6,7 @@ import {
   initializeMocks,
 } from '../../../../../testUtils';
 import editorStore from '../../../../data/store';
+import { EditorContextProvider } from '../../../../EditorContext';
 import * as hooks from './hooks';
 import SelectTypeModal from '.';
 
@@ -18,7 +19,13 @@ describe('SelectTypeModal', () => {
     const mockSelect = jest.fn();
     jest.spyOn(hooks, 'onSelect').mockImplementation(mockSelect);
     // This is a new-style test, unlike most of the old snapshot-based editor tests.
-    render(<Provider store={editorStore}><SelectTypeModal onClose={mockClose} /></Provider>);
+    render(
+      <EditorContextProvider fullScreen={false} learningContextId="course-v1:Org+COURSE+RUN">
+        <Provider store={editorStore}>
+          <SelectTypeModal onClose={mockClose} />
+        </Provider>
+      </EditorContextProvider>,
+    );
 
     // First we see the menu of problem types:
     expect(await screen.findByRole('button', { name: 'Numerical input' })).toBeInTheDocument();
